More re-factoring around Link pre-processing.  WORK IN PROGRESS...
