+++ src/main/scala/uk/gov/ons/bi/dataload/ubrn/LinksPreprocessor.scala
@@ -2,14 +2,9 @@ package uk.gov.ons.bi.dataload.ubrn
 
 import com.google.inject.Singleton
 import org.apache.spark.SparkContext
-import org.apache.spark.sql.functions._
-import org.apache.spark.sql.{DataFrame, SQLContext}
-import org.joda.time.DateTime
-import org.joda.time.format.DateTimeFormat
-import uk.gov.ons.bi.dataload.reader.{LinkJsonReader, PreviousLinksReader}
+import org.apache.spark.sql.DataFrame
+import uk.gov.ons.bi.dataload.reader.{LinkJsonReader, PreviousLinkStore}
 import uk.gov.ons.bi.dataload.utils.AppConfig
-
-import scala.util.{Success, Try}
 
 /**
   * Created by websc on 03/03/2017.
@@ -18,9 +13,8 @@ import scala.util.{Success, Try}
 @Singleton
 class LinksPreprocessor(sc: SparkContext) {
 
-
   def getNewLinksDataFromJson(reader: LinkJsonReader, appConfig: AppConfig): DataFrame = {
-    // Get source/target directories
+    // get source/target directories
     val linksDataConfig = appConfig.LinksDataConfig
     val dataDir = linksDataConfig.dir
     val jsonfile = linksDataConfig.json
@@ -28,26 +22,6 @@ class LinksPreprocessor(sc: SparkContext) {
 
     // Load the JSON links data
     reader.readFromSourceFile(jsonFilePath)
-  }
-
-  def writeAsPrevLinks(appConfig: AppConfig, df: DataFrame, timestamped: Boolean = false) = {
-    // Use timestamp as YYYYMMDD
-    val ts = if (timestamped) {
-      val fmt = DateTimeFormat.forPattern("yyyyMMddHHmm")
-
-      val now = DateTime.now()
-      now.toString(fmt)
-    }
-    else ""
-
-    // Parquet file locations from configuration (or runtime params)
-    val appDataConfig = appConfig.AppDataConfig
-    val linksFile = appDataConfig.links
-    val prevDir = appDataConfig.prevDir
-    val prevLinksFile = s"$prevDir/$ts/$linksFile"
-
-    // We will also write a copy of the preprocessed Links data to the "previous" dir
-    df.write.mode("overwrite").parquet(prevLinksFile)
   }
 
   def loadAndPreprocessLinks(appConfig: AppConfig) = {
@@ -68,27 +42,26 @@ class LinksPreprocessor(sc: SparkContext) {
     val newLinksFileParquetPath = s"$workingDir/$linksFile"
 
     // Get previous links
-    val previousLinksReader = new PreviousLinksReader(sc)
-    val prevLinks = previousLinksReader.readFromSourceFile(prevLinksFileParquetPath)
+    val previousLinkStore = new PreviousLinkStore(sc)
+    val prevLinks = previousLinkStore.readFromSourceFile(prevLinksFileParquetPath)
     prevLinks.cache()
 
     // Initialise LinkMatcher
     val matcher = new LinkMatcher(sc)
 
-    // Get easy matches (CH=CH, or no CH but other contents same)
+    // Get easy matches first (CH=CH, or no CH but other contents same)
     val (simpleChMatches, unmatched1) = matcher.simpleCompaniesHouseMatches(newLinks, prevLinks)
 
     // ----------------
     // CALL EXTRA MATCHING RULES HERE ...
     // ----------------
-    // Only need to try to match records that we have not already matched i.e. unmatched1
+    // Only try to match records that we have not already matched
     val (otherMatches, unmatched2) = matcher.matchedOnOtherRules(unmatched1, prevLinks)
 
     // Finally we should have :
-    // one sub-set of new links that we have matched and now have a UBRNs
+    // - one sub-set of new links that we have matched, so they now have a UBRN:
     val withOldUbrn: DataFrame = matcher.combineLinksToSave(simpleChMatches, otherMatches)
-
-    // and one sub-set of new links that we could not match and need new UBRN:
+    // - and one sub-set of new links that we could not match, so they need new UBRN:
     val needUbrn: DataFrame = unmatched2
 
     // ------------------------------
@@ -118,10 +91,10 @@ class LinksPreprocessor(sc: SparkContext) {
 
     // We will also write a copy of the new preprocessed Links data to the "previous" dir:
     // 1. As e.g. LINKS_Output.parquet so we can easily pick it up next time
-    writeAsPrevLinks(appConfig, linksToSave)
+    previousLinkStore.writeAsPrevLinks(appConfig, linksToSave)
 
     // 2. As e.g. 201703081145/LINKS_Output.parquet so it does not get over-written later
-    writeAsPrevLinks(appConfig, linksToSave, true)
+    previousLinkStore.writeAsPrevLinks(appConfig, linksToSave, true)
 
     // Clear cached data we no longer need
     linksToSave.unpersist()
